Falta corregir errores y tests

+++ src/main/java/Controlador/Llamada.java
@@ -5,9 +5,9 @@ import java.util.Calendar;
 public class Llamada {
     private int telefono;
     private Calendar fechaLlamada;
-    private int duracion;
+    private float duracion;
 
-    public Llamada(int telefono, Calendar fechaLlamada, int duracion){
+    public Llamada(int telefono, Calendar fechaLlamada, float duracion){
         this.duracion = duracion;
         this.fechaLlamada = fechaLlamada;
         this.telefono = telefono;
